refactor(routes): consolidate backend routing using Express Router pattern
Reorganized all route files to consistently use the Express Router pattern.
Separated protected and unprotected routes and registered them in index.ts
to improve maintainability and scalability.
