+++ Server/src/index.ts
@@ -3,15 +3,17 @@ dotenv.config();
 import express from 'express';
 import config from 'config';
 import mongoose from "mongoose";
-import userRoutes from './routes/auth';
+import authRoutes from './routes/auth';
 import cookieParser from 'cookie-parser';
 import cors from "cors";
 import {Request, Response, NextFunction} from "express";
 import commentRoutes from "./routes/comment";
-import authCheck, {authMiddleware, checkTokenExpiration} from "./middlewares/auth";
+import  {authMiddleware, checkTokenExpiration} from "./middlewares/auth";
 import postRoutes from "./routes/Post";
 import profileRoutes from "./routes/Profile";
 import messageRoutes from "./routes/Message";
+import protectedRoutes from "./routes/protectedAuth";
+import protectedPost from "./routes/protectedPost";
 const app = express();
 app.use(express.json({limit: '10mb'}));
 app.use(cookieParser());
@@ -36,8 +38,10 @@ app.get("/", (req: Request, res: Response) => {
 app.use(express.urlencoded({extended: true, limit: '10mb'}));
 
 //routes
-userRoutes(app, authMiddleware);
-postRoutes(app, authMiddleware);
+app.use("/Auth", authRoutes);
+app.use("/Auth", authMiddleware,protectedRoutes);
+app.use("/Post", postRoutes);
+app.use("/Post", authMiddleware,protectedPost);
 app.use("/Comment", authMiddleware, commentRoutes);
 app.use("/Profile", authMiddleware, profileRoutes);
 app.use("/Message", authMiddleware, messageRoutes);
