+++ Server/src/routes/protectedAuth.ts
@@ -1,0 +1,12 @@
+import {Request, Response,Router} from "express";
+import {getAllFollowings, getSendersForCurrentUser} from "../controllers/User";
+const protectedRoutes = Router();
+protectedRoutes.get("/following/:id", (req: Request, res: Response) => {
+    getAllFollowings(req,res);
+})
+protectedRoutes.get("/sentMessages", (req: Request, res: Response) => {
+    getSendersForCurrentUser(req,res);
+
+})
+
+export  default  protectedRoutes;
